Create the value of integer variables like ${42} with `int` not with `long`.
Update issue 834
Status: Done

+++ src/robot/variables/variables.py
@@ -110,7 +110,7 @@ class Variables(utils.NormalizedDict):
             raise ValueError
         base = self._normalize(name)[2:-1]
         try:
-            return long(base)
+            return int(base)
         except ValueError:
             return float(base)
